+++ word_search_generator/utils.py
@@ -17,9 +17,9 @@ def cleanup_input(words: str) -> set:
     Returns:
         set: Words to be placed in the puzzle.
     """
-    if type(words) != str:
+    if not isinstance(words, str):
         raise TypeError(
-            "Must be a string of words separated by spaces, commas, or new lines"
+            "Words must be a string separated by spaces, commas, or new lines"
         )
     # remove new lines
     words = words.replace("\n", ",")
@@ -28,20 +28,17 @@ def cleanup_input(words: str) -> set:
     # iterate through all words and pick first set that match criteria
     word_list = set()
     for word in words:
-        if len(word_list) >= config.max_puzzle_words:
+        if len(word_list) > config.max_puzzle_words:
             break
-        if len(word) > 1 and not check_for_punctuation(word):
+        if len(word) > 1 and not contains_punctuation(word):
             word_list.add(word.upper())
     # if no words were left raise exception
     if not word_list:
-        raise ValueError(
-            "No valid words. Use words that don't have punctuation and are longer than one-character."
-        )
-    # return words to be placed in the puzzle
+        raise ValueError("Use words longer than one-character and without punctuation.")
     return word_list
 
 
-def check_for_punctuation(word):
+def contains_punctuation(word):
     """Check to see if puncuation is present in the provided string."""
     return any([True if c in string.punctuation else False for c in word])
 
@@ -64,18 +61,35 @@ def stringify(puzzle: list, tabs: bool = False):
 
 
 def replace_right(s: str, target: str, replacement: str, replacements: int = 1) -> str:
+    """Replace `target` with `replacement` from the right size of the string.
+
+    Args:
+        s (str): A string to do replacement on.
+        target (str): String to be replaced.
+        replacement (str): New string to insert.
+        replacements (int, optional): Number of times replace. Defaults to 1.
+
+    Returns:
+        str: Updated string with replacement(s).
+    """
     return replacement.join(s.rsplit(target, replacements))
 
 
 def get_level_dirs_str(level: int) -> str:
+    """Return all pozzible directions for specified level."""
     return replace_right(", ".join(config.level_dirs[level]), ", ", ", and ")
 
 
 def get_word_list_str(key: dict) -> str:
+    """Return all placed puzzle words as a list."""
     return ", ".join([k for k in sorted(key.keys())])
 
 
 def get_answer_key_str(key: dict) -> str:
+    """Retunr a easy to read answer key for display/export."""
     return ", ".join(
-        [f"{k} {key[k]['dir']} @ {key[k]['start']}" for k in sorted(key.keys())]
+        [
+            f"{k} {key[k]['dir']} @ {(key[k]['start'][0] + 1, key[k]['start'][1] + 1)}"
+            for k in sorted(key.keys())
+        ]
     )